Added intro screen and end screen
Moved the contents of index.html to a new file, experiment.html. index.html is now an intro screen that introduces the experiment to the user as well as shows an example.

Additionally, after the final test, the entire screen changes to display a thank you message and prompt the user to submit their answer